change different types for price to unified f64

+++ src/model/bundle.rs
@@ -49,7 +49,7 @@ pub struct Bundle {
     /// Bundle description.
     pub description: String,
     /// Price in the default currency.
-    pub price: i32,
+    pub price: f64,
     /// Price in different currencies.
     pub currency_prices: CurrencyPrices,
     /// Whether the bundle is hidden.
